refactor: Lotto.test 에서 LottoManager.test로 분리
- 구입 금액에 해당하는 로또 장수를 구한다.
- 로또 장수에 따라 여러 장 발행한다.
